Resolved more comments
